requirement.txt
pandas
pandasai
streamlit
langchain_community
langchain
openpyxl
langchain_google_genai
matplotlib 
seaborn 
scipy
python-dotenv